requirements.txt
pillow>=10.0.0
python-dateutil>=2.8.2
RPi.GPIO>=0.7.1
fake-rpi>=0.7.1
tzdata>=2023.3
requests>=2.31.0

